Fixed issue with consent ignoring replacing the script tag inside script
With WordPress 6.2, Gutenberg began inserting scripts into the
JavaScript logic. Because of this change, our logic with ignoring
consent on selected scripts was working incorrectly, breaking the
JavaScript and causing editor to throw the error.

With this commit we ensure we're modifying the real script tags and not
some content inside the extra inline scripts.